test:cleanup pytest config
