+++ setup.cfg
@@ -16,10 +16,9 @@ split_before_named_assigns=False
 split_before_closing_bracket=False
 
 [tool:pytest]
-python_files = tests_*.py
-testpaths = tests
-addopts = -v --tb=short
-usefixtures = pretest_posttest
+python_files=tests_*.py
+testpaths=tests
+addopts=-v --tb=short -rxs --durations=0 --durations-min=0.1
 
 [metadata]
 name = tqdm
