+++ tests/tests_itertools.py
@@ -4,7 +4,6 @@ Tests for `tqdm.contrib.itertools`.
 import itertools
 
 from tqdm.contrib.itertools import product
-from .tests_tqdm import pretest_posttest  # NOQA, pylint: disable=unused-import
 from .tests_tqdm import StringIO, closing
