+++ tests/tests_keras.py
@@ -1,7 +1,6 @@
 from __future__ import division
 
 from tqdm import tqdm
-from .tests_tqdm import pretest_posttest  # NOQA, pylint: disable=unused-import
 from .tests_tqdm import importorskip, StringIO, closing
