+++ tests/tests_perf.py
@@ -11,7 +11,6 @@ except ImportError:
 import sys
 
 from tqdm import tqdm, trange
-from .tests_tqdm import pretest_posttest  # NOQA, pylint: disable=unused-import
 from .tests_tqdm import importorskip, skip, StringIO, closing, _range, \
     patch_lock
