+++ tox.ini
@@ -26,7 +26,7 @@ deps =
     {[coverage]deps}
     codecov
 commands =
-    pytest --durations=10 --cov=tqdm --cov-report=xml --cov-report=term -k "not tests_perf"
+    pytest --cov=tqdm --cov-report=xml --cov-report=term -k "not tests_perf"
     - coveralls
     codecov
     - codacy report -r coverage.xml
@@ -57,7 +57,7 @@ deps = {[extra]deps}
 
 [testenv:perf]
 deps = pytest
-commands = pytest --durations=0 -k tests_perf
+commands = pytest -k tests_perf
 
 [testenv:flake8]
 deps = flake8
